# contracts/openapi.yaml
openapi: 3.1.0
info:
  title: Food Court QR Ordering Platform API
  version: 1.0.0-pilot-STALE
  description: |
    # ⚠ THIS DOCUMENT IS STALE. DO NOT GENERATE A CLIENT FROM IT.

    Superseded by `docs/prd.md` v7.0 and by the running code. It is kept only so
    the divergence is visible; every disagreement below resolves in favour of
    the implementation.

    The warning lives here, in the file itself, rather than only in the PRD.
    Somebody who opens this directly — to generate a client, to check a field
    name, to write an integration — would otherwise get no warning at all, and
    a stale contract read in good faith produces confidently wrong code.

    ## What is known to be wrong

    | Here | Actually |
    | --- | --- |
    | `/qr/{token}` returns a `table` object | It resolves to a FOOD COURT. The table concept was removed — PRD §2.1 |
    | Order status `COMPLETED` | `COLLECTED`. Cooked and handed over are different events — PRD §7.1 |
    | Payment status `SUCCESS` | `AUTHORIZED` and `CAPTURED` are separate states — PRD §8.1 |
    | Menu item `is_available` boolean | `status` / `availability` / `inventory_mode`, three concepts — PRD §6 |
    | `POST /kds/orders/{id}/complete` | `/collect` |
    | no session or OTP endpoints | `/auth/otp/request`, `/auth/otp/verify` exist — PRD §11.2 |
    | no payment endpoints | `/orders/{id}/payment-intent`, `/webhooks/payments`, `/orders/{id}/payment` exist |
...

...
    script produces types that contradict the server.

...

...

...

...

...
servers:
...
security:
  - sessionToken: []
paths:
...
  /checkout/validate:
    post:
      tags: [Checkout]
      operationId: validateCheckout
      summary: Revalidate everything and return the authoritative bill
      description: |
        PRD CUS-CHECK-01/02. Revalidates vendor availability, **device health**,
        item stock and pricing, then returns the bill the customer will be
        charged, itemised. The client never computes the payable amount.
        Call this immediately before creating the order.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cartId]
              properties:
                cartId: { type: string, format: uuid }
                applyCreditId:
                  type: string
                  format: uuid
                  nullable: true
                  description: Optional platform credit to apply (PRD REF-05).
      responses:
        '200':
          description: Checkout is valid; bill is authoritative for 120 seconds.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CheckoutQuote' }
        '409':
          description: |
            Blocking condition. `code` is one of `VENDOR_CLOSED`,
            `VENDOR_DEVICE_OFFLINE`, `ITEM_UNAVAILABLE`, `PRICE_CHANGED`,
            `BELOW_MINIMUM_ORDER`, `CART_EXPIRED`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
  /orders:
    post:
      tags: [Checkout]
      operationId: createOrder
      summary: Create an order and its payment intent
      description: |
        PRD PAY-01/02, API-01. Requires `Idempotency-Key`. Creates the order in
        `CREATED`, snapshots the fee rule, settlement mode and tax model, then
        creates the provider payment intent and moves the order to
        `PAYMENT_PENDING`. Replaying the same key returns the original order and
        payment intent with `200`.

        This endpoint never confirms payment. Confirmation arrives only through
        the verified provider webhook.
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [cartId, quoteId, customerPhone]
              properties:
                cartId: { type: string, format: uuid }
                quoteId:
                  {
                    type: string,
                    format: uuid,
                    description: From /checkout/validate. Binds the price.,
                  }
                customerPhone:
                  type: string
                  pattern: '^\+[1-9][0-9]{7,14}$'
                  description: |
                    Captured, not verified. PRD §8.2 — payment success is the
                    implicit verification event. No OTP in the pilot.
                whatsappOptIn: { type: boolean, default: false }
                applyCreditId: { type: string, format: uuid, nullable: true }
      responses:
        '201':
          description: Order created, payment intent issued
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderCreated' }
        '200':
          description: Idempotent replay — the original order is returned unchanged.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/OrderCreated' }
        '400':
          description: Missing `Idempotency-Key`.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }
        '409':
          description: Quote expired or checkout no longer valid. Re-run /checkout/validate.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ==================================================================
  # PAYMENT
  # ==================================================================
...
  /orders/{orderId}:
    get:
      tags: [Order (customer)]
      operationId: getOrder
      description: |
        PRD CUS-TRK-04. Clients treat socket events as hints and re-fetch this
        endpoint for authoritative state. Never derive state from event ordering.
      parameters: [{ $ref: '#/components/parameters/OrderId' }]
      responses:
        '200':
          description: Order state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '404': { $ref: '#/components/responses/NotFound' }
  /orders/{orderId}/cancel:
    post:
      tags: [Order (customer)]
      operationId: customerCancelOrder
      summary: Customer-initiated cancellation, available only when offered
      description: |
        PRD ESC-03. Enabled only once the escalation ladder has reached 180s and
        the order is `DISPATCH_FAILED`. One tap, no support contact, refund
        initiated in the same request. At any other time this returns `409`.
      parameters:
        - { $ref: '#/components/parameters/OrderId' }
        - { $ref: '#/components/parameters/IdempotencyKey' }
      responses:
        '200':
          description: Cancelled and refund initiated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Order' }
        '409':
          description: Cancellation is not currently offered for this order.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Problem' }

  # ==================================================================
  # KDS — VENDOR DEVICE
  # ==================================================================
...
  /kds/orders:
    get:
      tags: [KDS]
      operationId: kdsListOrders
      security: [{ deviceToken: [] }]
      parameters:
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/OrderStatus' }
      responses:
        '200':
          description: Live queue for this vendor
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/KdsOrder' }
...
  /kds/vendor/ordering:
    patch:
      tags: [KDS]
      operationId: setVendorOrdering
      security: [{ deviceToken: [] }]
      description: PRD KDS-STK-02. Blocks new orders immediately; in-flight orders unaffected.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [acceptingOrders]
              properties:
                acceptingOrders: { type: boolean }
                reopenAt: { type: string, format: date-time, nullable: true }
      responses:
        '200': { description: Updated }

  # ==================================================================
  # MANAGER
  # ==================================================================
  /manager/orders:
    get:
      tags: [Manager]
      operationId: managerListOrders
      description: PRD MGR-01. Consolidated live feed across all vendors in the court.
      parameters:
        - name: foodCourtId
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: status
          in: query
          schema:
            type: array
            items: { $ref: '#/components/schemas/OrderStatus' }
        - name: vendorId
          in: query
          schema: { type: string, format: uuid }
        - { $ref: '#/components/parameters/Limit' }
        - { $ref: '#/components/parameters/Cursor' }
      responses:
        '200':
          description: Paged order feed
          content:
            application/json:
              schema:
                type: object
                properties:
                  orders:
                    type: array
                    items: { $ref: '#/components/schemas/Order' }
                  nextCursor: { type: string, nullable: true }
...
components:
  securitySchemes:
    sessionToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Short-lived customer session token issued by POST /sessions.
    deviceToken:
      type: http
      scheme: bearer
      description: KDS or printer device token. Scoped to one vendor.
    staffToken:
      type: http
      scheme: bearer
      description: Manager, ops, finance and admin. Role and tenant scope are claims.
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      description: |
        Client-generated, unique per attempt. Backed by a unique database
        constraint. Replay returns the original result. PRD API-01.
      schema: { type: string, minLength: 16, maxLength: 128 }
...
    OrderId:
      name: orderId
      in: path
      required: true
      schema: { type: string, format: uuid }
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    Cursor:
      name: cursor
      in: query
      schema: { type: string }
  responses:
    NotFound:
      description: |
        Not found, or not visible to this tenant. Cross-tenant reads return 404
        rather than 403 so existence is not leaked (PRD RBAC-01).
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Problem' }
...
  schemas:
    Problem:
      type: object
      required: [code, message, correlationId]
      properties:
        code:
          type: string
          description: |
            Stable machine-readable error code. Clients branch on this, never on
            the message. See TDD §18 for the full catalogue.
          examples:
            [VENDOR_DEVICE_OFFLINE, INVALID_TRANSITION, QUOTE_EXPIRED, IDEMPOTENCY_KEY_REQUIRED]
        message: { type: string, description: Human-readable, not for display to customers. }
        detail: { type: string, nullable: true }
        correlationId: { type: string }
        fields:
          type: object
          nullable: true
          additionalProperties: { type: string }
...
    OrderStatus:
      type: string
      enum:
        [
          CREATED,
          PAYMENT_PENDING,
          PAYMENT_CONFIRMED,
          DISPATCHED,
          ACKNOWLEDGED,
          PREPARING,
          READY,
          COMPLETED,
          PAYMENT_FAILED,
          DISPATCH_FAILED,
          REJECTED,
          CANCELLED,
          REFUND_PENDING,
          REFUNDED,
          REFUND_FAILED,
          RECONCILIATION_REQUIRED,
        ]
...
    CheckoutQuote:
      type: object
      description: |
        The authoritative bill. Every line the customer will pay is named — no
        undisclosed platform fee (PRD CUS-CHECK-02). Binding until `expiresAt`.
      required: [quoteId, expiresAt, totalPayablePaise]
      properties:
        quoteId: { type: string, format: uuid }
        expiresAt: { type: string, format: date-time }
        subtotalPaise: { $ref: '#/components/schemas/Money' }
        foodTaxPaise: { $ref: '#/components/schemas/Money' }
        customerFeePaise: { $ref: '#/components/schemas/Money' }
        customerFeeTaxPaise: { $ref: '#/components/schemas/Money' }
        creditAppliedPaise: { $ref: '#/components/schemas/Money' }
        totalPayablePaise: { $ref: '#/components/schemas/Money' }
        lines:
          type: array
          description: Display-ready breakdown, already labelled for the UI.
          items:
            type: object
            properties:
              label: { type: string }
              amountPaise: { type: integer, format: int64 }
              kind: { type: string, enum: [SUBTOTAL, TAX, FEE, DISCOUNT, CREDIT, TOTAL] }
    OrderCreated:
      type: object
      properties:
        order: { $ref: '#/components/schemas/Order' }
        payment:
          type: object
          properties:
            paymentId: { type: string, format: uuid }
            provider: { type: string }
            providerOrderRef: { type: string }
            amountPaise: { $ref: '#/components/schemas/Money' }
            checkoutPayload:
              type: object
              description: Opaque provider payload the client hands to the checkout SDK.
              additionalProperties: true
    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        publicOrderNumber: { type: string }
        status: { $ref: '#/components/schemas/OrderStatus' }
        foodCourtId: { type: string, format: uuid }
        vendorId: { type: string, format: uuid }
        vendorName: { type: string }
        tableLabel: { type: string }
        items:
          type: array
          items:
            type: object
            properties:
              name: { type: string }
              quantity: { type: integer }
              lineTotalPaise: { $ref: '#/components/schemas/Money' }
              options:
                type: array
                items: { type: string }
              instructions: { type: string, nullable: true }
              isRejected: { type: boolean }
        totalPayablePaise: { $ref: '#/components/schemas/Money' }
        rejectionReason:
          oneOf:
            - { $ref: '#/components/schemas/RejectionReason' }
            - { type: 'null' }
        refund:
          oneOf:
            - { $ref: '#/components/schemas/Refund' }
            - { type: 'null' }
        creditIssued:
          oneOf:
            - { $ref: '#/components/schemas/PlatformCredit' }
            - { type: 'null' }
        cancellationOffered:
          type: boolean
          description: True once the escalation ladder reaches 180s (PRD ESC-03).
        estimatedReadyAt: { type: string, format: date-time, nullable: true }
        timeline:
          type: array
          items:
            type: object
            properties:
              status: { $ref: '#/components/schemas/OrderStatus' }
              at: { type: string, format: date-time }
              actorType: { type: string }
        createdAt: { type: string, format: date-time }
...